+++ libs/core/src/lib/api/index.ts
@@ -2,5 +2,4 @@ export * from './account.service';
 export * from './sdk.service';
 export * from './sso.service';
 export * from './user.service';
-export * from './users.service';
 export * from './zone.service';