+++ libs/core/src/lib/models/index.ts
@@ -4,6 +4,5 @@ export * from './magic.model';
 export * from './permissions.model';
 export * from './region.model';
 export * from './user.model';
-export * from './users.model';
 export * from './welcome.model';
 export * from './zone.model';